hotfix: Fix infinite cursor chat issue by partially reverting "reactive context menu overrides (#2697)" (#2775) (#2777)
Hotfixed from f185edcb76b5da78a8d40ce129861643e3d302d0

When we made context menu overrides, we introduced two new issues:
1. the context menu on the main app now updates much more frequently
than it should
2. every time it updates, it adds a new 'cursor chat' entry to the menu

This reverts the part of that change that made the context menu
reactive. This is the quick fix for us to hotfix, but i'm going to
follow this up by restoring that functionality without those issues.